Fix hang if no backends are available
This patch prevents gupnp-dlna from hanging when no meta data extracting
backends are available.  The problem occurs as the existing code passes a
value of 0 into g_once_init_leave if a backend cannot be loaded.  You're
not allowed to do this and as a result load_metadata_backend hangs the
next time it is called.  The patch fixes the problem by using 1 and 2 to
indicate failure and success, respectively.

https://bugzilla.gnome.org/show_bug.cgi?id=707909